+++ CHANGELOG.md
@@ -1,1 +1,6 @@
+07-08-2022 SpreadSheetDiff version 0.1.1 released: 
+
+* New Feature: highlighting differences in resulting exel files (background color, bold text style)
+* New command line options for highlighting feature
+
 31-07-2022 SpreadSheetDiff version 0.1.0 released